no warnings

+++ src/2opt.cpp
@@ -12,7 +12,7 @@ void optimize(std::vector<Point*> & points, NearMatrix& matrix){
   bool improving;
   do{
     improving = false;
-    for(int i=0; i<matrix.size(); ++i){
+    for(int i=0; i<(int)matrix.size(); ++i){
       if(findSwap(matrix, points[i]))
 	improving = true;
     }
@@ -22,7 +22,7 @@ void optimize(std::vector<Point*> & points, NearMatrix& matrix){
 bool findSwap(NearMatrix& nearMat, Point *t1){
   Point *t2 = t1->next;//index för staden efter t1 just nu
 
-  for(int i=1; i<nearMat[t2->i].size(); ++i){ //iterera över städer nära t2
+  for(int i=1; i<(int)nearMat[t2->i].size(); ++i){ //iterera över städer nära t2
     Point *t3 = nearMat[t2->i][i].point;
     Point *t4 = t3->next;
